easypg: allow to assert on content diffs instead of just on snapshots
In real-world testcases, we often have a large setup phase, where it is
appropriate to assert on a full snapshot at the end. But then the actual
test has a number of small steps where the DB changes only a bit, and we
want to assert on this diff. Until now, applications had sequences of
fixture files for each test, but that obscures the actual diff
needlessly. Also, it hinders tests with identical operations, but
different setups, from reusing fixtures.