easypg: add URLFrom()
I was initially tempted to have all the getenv calls inline inside the
function, so that only a variable name prefix would have to be provided
(e.g. `URLFrom("TENSO_DB_")` and such), but for the sake of clarify and
greppability, I want to have it so that every environment variable that
a program recognizes appears in full in the program's own code.