jobloop: add InitialDelay option
In Limes, when a new service type is added, I need the
CheckConsistencyJob to run pretty much immediately after startup to
create the respective service and resource entries. But since its
interval is set to 1h (which makes sense for normal operations), we will
have the DB in an inconsistent state for an hour after deployment.

With this change, I can give it a short InitialDelay of 10s or so to
have the inconsistencies rectified more quickly.