jobloop: add WithLabel option
This allows running multiple instances of the same Job in parallel while
keeping the individual instances separated. I need this in Limes where
we run one scrape job per service type. I cannot just use a
ProducerConsumerJob because the overall scrape job is not
ConcurrencySafe. Only the individual job instances are safe to run
concurrently to each other because they are all confined to their own
service type and thus cannot interfere with each other.

This interface is my second attempt at covering this usecase, and is
much less invasive than the first attempt. The major change is that
pretty much everything needs to have a jobConfig now, which also means
that the public API needs to change to be able to pass the WithLabel
option into ProcessOne. Thankfully, that's a backwards-compatible change
because only a variadic argument is added.