audittools: add type Auditor
When building the OpenStack service blueprint repo, I noticed that there
is still a lot of repetitive boilerplate that is necessary to set up an
audit event submitter. This commit adds the weighted average of all
existing users into the library:

https://github.com/sapcc/andromeda/blob/main/middlewares/audit.go
https://github.com/sapcc/archer/blob/main/internal/middlewares/audit.go
https://github.com/sapcc/castellum/blob/master/internal/api/audit.go
https://github.com/sapcc/limes/blob/master/internal/api/audit.go
https://github.com/sapcc/keppel/blob/master/internal/keppel/auditor.go

The various logfile-only options in Castellum, Limes and Keppel are
replaced with type NullAuditor that logs into the debug log, to make it
clear that this usecase is not suitable for productive deployments.
The option to log to a non-debug log hails back to the ancient times
when Hermes was not deployed everywhere yet.

The Auditor interface is added as a dependency injection slot for
automated tests. Keppel donates its MockAuditor type which ought to be
useful in other applications, too.