audittools: turn NewEvent into a method of EventParameters
This is mostly for optics. Application developers are not supposed to
use it, but a free-standing function called "New" is just very
prominent. By turning it into a method with a more specific (and
clunkier) name, it will be more easily recognized as low-level plumbing.

Regardless, I chose to leave this method in the public API because it is
useful for foreign implementors of the Auditor interface (e.g. mock
auditors that need to work in a different way for application-specific
reasons).