easypg: add management of a test DB server
This replaces the shell script `./testing/with-postgres-db.sh`.
The new test server is completely separate from the one managed by the
script (it uses a different data directory and also a different port
number), so it can be run in parallel with the old script until all
usage has been ported over. Then we can remove the shell script from
go-makefile-maker.