add new package httptest
This is intended as a replacement for type assert.HTTPRequest,
which does not play well with Ginkgo/Gomega. The big problem is that it
wants to log several lines with t.Errorf() and t.Logf(). But in Ginkgo,
Errorf() is aliased to Fatalf(), so we will be missing important log
lines after the first error.