audittools, respondwith: unshare GenerateUUID
I suggested to Valentin to share this function to avoid a dependency
in respondwith on audittools. But I was not aware that the function
implementation is literally a one-liner.

Since we need to retain GenerateUUID() as an exported symbol in
audittools for backwards compatibility, the complications of having a
shared implementation are so large that I find it easier to just copy
that one line in respondwith.