respondwith: add func CustomStatus
Over in Limes, there is lots of code duplication in the commitment API
that I want to reduce, but client error responses are making this
difficult. For example: (some args elided for brevity)

    var (
      resourceID                db.ProjectResourceID
      azResourceID              db.ProjectAZResourceID
      resourceAllowsCommitments bool
    )
    err := p.DB.QueryRow(findProjectAZResourceIDByLocationQuery, args...).
      Scan(&resourceID, &azResourceID, &resourceAllowsCommitments)
    if respondwith.ErrorText(w, err) {
      return
    }
    if !resourceAllowsCommitments {
      msg := fmt.Sprintf("resource %s/%s is not enabled in this project", args...)
      http.Error(w, msg, http.StatusUnprocessableEntity)
      return
    }

Splitting this into a helper function would require two error code
paths, one for generic server errors from the DB query failing, and
another for 422 errors from the !resourceAllowsCommitments check.
We would have something like:

    resourceID, azResourceID, ok := p.findProjectAZResourceIDByLocation(w, args)
    if !ok {
      return
    }

And the implementation of that helper function would need to take an
http.ResponseWriter. That's fine, I guess, and matches existing
precedent (see e.g. go-bits/gopherpolicy.Token.Require()). But I don't
like that all these helpers need to take a ResponseWriter and thus are
able to mess with the response in ways that they should not be allowed
to do.

This commit adds an alternate design where the helper function can
return an error that is annotated with a specific response status code
that respondwith.ErrorText() will use instead of the default of 500.