httptest: change Handler.RespondTo() to return a custom Response type
This is a breaking change, but I don't know of any users of this API as
of now. I devised it for use in the openstack-blueprint together with
Ginkgo, but I turned out to not like Ginkgo at all and thus removed its
only known use in Surveyor very quickly.

By returning a custom type in RespondTo(), we can add additional
assertion functions on that type, starting in the next commit.

The changes to the test file showcase how to interact with the new type.