add syncext.Semaphore
This already kind of existed within the implementation of
httpext.LimitConcurrentRequestsMiddleware(), but I want to use it in
Keppel and Limes, too. It's not a lot of code and I could have just
copy-pasted it, but giving it a proper name should make the code using
it downstream more readable.